Add ClientInfo() server method
This also adds a proper struct to hold the response data. Documentation
of the fields is a mix of the server query manual and experiments with
an actual server. I could not figure out the meaning or data type for
all fields, most notably `client_badges` and `client_meta_data` which I
even excluded from the struct definition.

Data for the tests is the incomplete data from the manual supplemented
by real query results. Fields like ip address and UUIDs have been
anonymized by replacing characters. I tried not to introduce characters
which were previously not used in the value.